Use the branch sdk to generate deep links.
