+++ package.json
@@ -28,6 +28,7 @@
     "@r/widgets": "~0.2.2",
     "autoprefixer": "6.3.6",
     "babel-polyfill": "^6.7.4",
+    "branch-sdk": "^2.11.0",
     "Base64": "1.0.0",
     "cluster": "^0.7.7",
     "event-tracker": "git://github.com/reddit/event-tracker.git#28bc2dd9ce3b80540fdf189e1b003f8264cd0d08",
