+++ src/app/reducers/smartBanner.js
@@ -7,6 +7,7 @@ export const DEFAULT = {
   showBanner: false,
   haveShownXPromo: false,
   xPromoShownUrl: null,
+  clickUrl: null,
 };
 
 export default function(state=DEFAULT, action={}) {
