+++ src/lib/smartBannerState.js
@@ -1,94 +1,10 @@
-import url from 'url';
-import cookies from 'js-cookie';
-
 import localStorageAvailable from './localStorageAvailable';
-import { getDevice, IOS_DEVICES, ANDROID } from 'lib/getDeviceFromState';
-import * as constants from 'app/constants';
-import features from 'app/featureFlags';
+import { generateBranchLink } from 'lib/branch';
 
 const TWO_WEEKS = 2 * 7 * 24 * 60 * 60 * 1000;
 
-const { USE_BRANCH } = constants.flags;
-
-// Get loid values either from the account state or the cookies.
-function getLoidValues(accounts) {
-  if (accounts.me) {
-    return {
-      loid: accounts.me.loid,
-      loidCreated: accounts.me.loidCreated,
-    };
-  }
-
-  const loid = cookies.get('loid');
-  const loidCreated = cookies.get('loidcreated');
-
-  return {
-    loid,
-    loidCreated,
-  };
-}
-
-export function getBranchLink(state, payload={}) {
-  const { user, accounts } = state;
-
-  const { loid, loidCreated } = getLoidValues(accounts);
-
-  let userName;
-  let userId;
-
-  const userAccount = user.loggedOut ? null : accounts[user.name];
-  if (userAccount) {
-    userName = userAccount.name;
-    userId = userAccount.id;
-  }
-
-  const basePayload = {
-    channel: 'mweb_branch',
-    feature: 'smartbanner',
-    campaign: 'xpromo_banner',
-    // We can use this space to fill "tags" which will populate on the
-    // branch dashboard and allow you sort/parse data. Optional/not required.
-    // tags: [ 'tag1', 'tag2' ],
-    // Pass in data you want to appear and pipe in the app,
-    // including user token or anything else!
-    '$og_redirect': window.location.href,
-    '$deeplink_path': window.location.href.split(window.location.host)[1],
-    mweb_loid: loid,
-    mweb_loid_created: loidCreated,
-    utm_source: 'mweb_branch',
-    utm_medium: 'smartbanner',
-    utm_name: 'xpromo_banner',
-    mweb_user_id36: userId,
-    mweb_user_name: userName,
-  };
-
-  return url.format({
-    protocol: 'https',
-    host: 'reddit.app.link',
-    pathname: '/',
-    query: {...basePayload, ...payload},
-  });
-}
-
 export function getDeepLink(state) {
-  const device = getDevice(state);
-
-  // See if we should use a Branch link
-  const feature = features.withContext({ state });
-  if (feature && feature.enabled(USE_BRANCH)) {
-    // just use the universal Branch link
-    return getBranchLink(state);
-  }
-
-  // Otherwise use a basic deep link
-
-  if (IOS_DEVICES.includes(device)) {
-    return constants.BANNER_URLS_DIRECT.IOS;
-  }
-
-  if (device === ANDROID) {
-    return constants.BANNER_URLS_DIRECT.ANDROID;
-  }
+  return generateBranchLink(state);
 }
 
 export function shouldShowBanner() {
